CLRadeonExtender: Fix compilation error on 32-bit systems.

+++ CLRX/amdbin/ROCmBinaries.h
@@ -48,7 +48,7 @@ enum : Flags {
 struct ROCmRegion
 {
     CString regionName; ///< region name
-    size_t size;    ///< data size
+    uint64_t size;    ///< data size
     uint64_t offset;     ///< data
     bool isKernel;
 };
